Move DocsInfo out of the components folder
